--- stripeTrueUp/function.json
@@ -1,10 +1,0 @@
-{
-  "bindings": [
-    {
-      "name": "myTimer",
-      "type": "timerTrigger",
-      "direction": "in",
-      "schedule": "0 0 * * * *"
-    }
-  ]
-}
